PB-722 : add file drag&drop overlay to 3D viewer
