+++ src/modules/map/components/cesium/CesiumMap.vue
@@ -165,8 +165,8 @@ provide('getViewer', () => viewer)
             ref="compassElement"
             class="position-absolute start-50 translate-middle-x cesium-compass"
         />
-        <slot />
     </div>
+    <slot />
 </template>
 
 <style lang="scss" scoped>
