+++ src/modules/map/components/openlayers/utils/useMapInteractions.composable.js
@@ -8,13 +8,10 @@ import LayerFeature from '@/api/features/LayerFeature.class'
 import LayerTypes from '@/api/layers/LayerTypes.enum'
 import { DRAWING_HIT_TOLERANCE } from '@/config/map.config'
 import { IS_TESTING_WITH_CYPRESS } from '@/config/staging.config'
+import useDragFileOverlay from '@/modules/map/components/common/useDragFileOverlay.composable'
 import { useDragBoxSelect } from '@/modules/map/components/openlayers/utils/useDragBoxSelect.composable'
-import { handleFileContent } from '@/modules/menu/components/advancedTools/ImportFile/utils'
 import { ClickInfo, ClickType } from '@/store/modules/map.store'
-import { normalizeExtent, OutOfBoundsError } from '@/utils/coordinates/coordinateUtils'
-import ErrorMessage from '@/utils/ErrorMessage.class'
-import { EmptyGPXError } from '@/utils/gpxUtils'
-import { EmptyKMLError } from '@/utils/kmlUtils'
+import { normalizeExtent } from '@/utils/coordinates/coordinateUtils'
 import log from '@/utils/logging'
 
 const dispatcher = {
@@ -66,12 +63,10 @@ export default function useMapInteractions(map) {
     })
 
     registerPointerEvents()
-    registerDragAndDropEvent()
     map.addInteraction(freeMouseWheelInteraction)
 
     onBeforeUnmount(() => {
         unregisterPointerEvents()
-        unregisterDragAndDropEvent()
     })
 
     /*
@@ -178,6 +173,7 @@ export default function useMapInteractions(map) {
                 pixelCoordinate: pixel,
                 features,
                 clickType: ClickType.LEFT_SINGLECLICK,
+                ...dispatcher,
             }),
         })
     }
@@ -191,6 +187,7 @@ export default function useMapInteractions(map) {
                 pixelCoordinate: event.pixel,
                 features: [],
                 clickType: ClickType.CONTEXTMENU,
+                ...dispatcher,
             }),
         })
     }
@@ -228,94 +225,5 @@ export default function useMapInteractions(map) {
         }
     }
 
-    function registerDragAndDropEvent() {
-        log.debug(`Register drag and drop events`)
-        const mapElement = map.getTargetElement()
-        mapElement.addEventListener('dragover', onDragOver)
-        mapElement.addEventListener('drop', onDrop)
-        mapElement.addEventListener('dragleave', onDragLeave)
-    }
-
-    function unregisterDragAndDropEvent() {
-        log.debug(`Unregister drag and drop events`)
-        const mapElement = map.getTargetElement()
-        mapElement.removeEventListener('dragover', onDragOver)
-        mapElement.removeEventListener('drop', onDrop)
-        mapElement.removeEventListener('dragleave', onDragLeave)
-    }
-
-    /**
-     * @param {File} file
-     * @returns {Promise<ArrayBuffer>}
-     */
-    function readFileContent(file) {
-        return new Promise((resolve, reject) => {
-            const reader = new FileReader()
-            reader.onload = (event) => resolve(event.target.result)
-            reader.onerror = (error) => reject(error)
-            // The file might be a KMZ file, which is a zip archive. Reading zip archive as text
-            // is asking for trouble therefore we use ArrayBuffer
-            reader.readAsArrayBuffer(file)
-        })
-    }
-
-    /**
-     * @param {File} file
-     * @returns {Promise<void>}
-     */
-    async function handleFile(file) {
-        try {
-            const fileContent = await readFileContent(file)
-            await handleFileContent(store, fileContent, file.name, file)
-        } catch (error) {
-            let errorKey
-            log.error(`Error loading file`, file.name, error)
-            if (error instanceof OutOfBoundsError) {
-                errorKey = 'imported_file_out_of_bounds'
-            } else if (error instanceof EmptyKMLError || error instanceof EmptyGPXError) {
-                errorKey = 'kml_gpx_file_empty'
-            } else {
-                errorKey = 'invalid_import_file_error'
-                log.error(`Failed to load file`, error)
-            }
-            store.dispatch('addErrors', { error: new ErrorMessage(errorKey, null), ...dispatcher })
-        }
-    }
-
-    function onDragOver(event) {
-        event.preventDefault()
-        store.dispatch('setShowDragAndDropOverlay', { showDragAndDropOverlay: true, ...dispatcher })
-    }
-
-    function onDrop(event) {
-        event.preventDefault()
-        store.dispatch('setShowDragAndDropOverlay', {
-            showDragAndDropOverlay: false,
-            ...dispatcher,
-        })
-
-        if (event.dataTransfer.items) {
-            // Use DataTransferItemList interface to access the file(s)
-            for (let i = 0; i < event.dataTransfer.items.length; i++) {
-                // If dropped items aren't files, reject them
-                if (event.dataTransfer.items[i].kind === 'file') {
-                    const file = event.dataTransfer.items[i].getAsFile()
-                    handleFile(file)
-                }
-            }
-        } else {
-            // Use DataTransfer interface to access the file(s)
-            for (let i = 0; i < event.dataTransfer.files.length; i++) {
-                const file = event.dataTransfer.files[i]
-                handleFile(file)
-            }
-        }
-    }
-
-    function onDragLeave() {
-        store.dispatch('setShowDragAndDropOverlay', {
-            showDragAndDropOverlay: false,
-            ...dispatcher,
-        })
-    }
+    useDragFileOverlay(map.getTargetElement())
 }
